+++ test/ui/workflow_list_page_test.dart
@@ -7,7 +7,16 @@ import 'package:flutter_test/flutter_test.dart';
 
 class _WorkflowRepository implements WorkflowRepository {
   @override
-  Future<List<WorkflowFieldOption>> getConditionFields(String definition) =>
+  Future<WorkflowDesign> addConditionBranch({
+    required String definition,
+    required String conditionStage,
+    required WorkflowStageType type,
+    required bool result,
+  }) async =>
+      throw UnimplementedError();
+  @override
+  Future<List<WorkflowFieldOption>> getConditionFields(String definition,
+          {String? beforeStage}) =>
       throw UnimplementedError();
 
   @override